issue #11 refactor: styled-component 분리
