+++ src/components/ToDoItem/styled.js
@@ -1,7 +1,4 @@
-import React from "react";
 import styled, { css } from "styled-components";
-import { MdDone, MdDelete } from "react-icons/md";
-import { useTodoDispatch } from "../TodoContext";
 
 const Remove = styled.div`
   display: flex;
@@ -63,22 +60,9 @@ const Text = styled.div`
     `}
 `;
 
-function TodoItem({ id, done, text }) {
-  const dispatch = useTodoDispatch();
-  const onToggle = () => dispatch({ type: "TOGGLE", id });
-  const onRemove = () => dispatch({ type: "REMOVE", id });
-
-  return (
-    <TodoItemBlock>
-      <CheckCircle done={done} onClick={onToggle}>
-        {done && <MdDone />}
-      </CheckCircle>
-      <Text done={done}>{text}</Text>
-      <Remove onClick={onRemove}>
-        <MdDelete />
-      </Remove>
-    </TodoItemBlock>
-  );
-}
-
-export default React.memo(TodoItem);
+export default {
+  Remove,
+  TodoItemBlock,
+  CheckCircle,
+  Text,
+};
