+++ src/components/ToDoList/styled.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const TodoListBlock = styled.div`
+  flex: 1;
+  padding: 20px 32px;
+  padding-bottom: 48px;
+  overflow-y: auto;
+`;
+
+export default {
+  TodoListBlock,
+};
